refactor: alterando a efeito no featured
